removing reth network api dependency from rpc types
