+++ crates/net/network-api/src/test_utils.rs
@@ -1,11 +1,10 @@
-use crate::{
-    EthProtocolInfo, NetworkError, NetworkInfo, NetworkStatus, PeerKind, Peers, PeersInfo,
-    ReputationChangeKind,
-};
+use crate::{NetworkError, NetworkInfo, PeerKind, Peers, PeersInfo, ReputationChangeKind};
 use async_trait::async_trait;
 use reth_eth_wire::{DisconnectReason, ProtocolVersion};
 use reth_primitives::{rpc::Chain::Mainnet, NodeRecord, PeerId};
 use std::net::{IpAddr, SocketAddr};
+
+use reth_rpc_types::{EthProtocolInfo, NetworkStatus};
 
 /// A type that implements all network trait that does nothing.
 ///
